% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_chen.R
\name{best_chen}
\alias{best_chen}
\title{best_chen}
\usage{
best_chen(auto_chen)
}
\arguments{
\item{auto_chen}{a model created by `auto_chen`}
}
\description{
This function is useful when a well-optimized combination of variables
is already known (by `auto_chen`) and the goal is to fit the model using only
those selected covariates
}
