% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.chenreg.R
\name{AIC.chenreg}
\alias{AIC.chenreg}
\title{BIC.chenreg}
\usage{
\method{AIC}{chenreg}(object, ...)
}
\arguments{
\item{object}{a fitted model created by `chen_reg`,
for which there exists the log-likelihood.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
BIC.chenreg
}
