% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chen_envlp.R
\name{chen_envlp}
\alias{chen_envlp}
\title{Simulated envelop for chen distribution}
\usage{
chen_envlp(z, b)
}
\arguments{
\item{z}{a model created by chen_reg.fit
a symbolic description of the model to be fitted}

\item{b}{Quantity of resamples}
}
\description{
Simulated envelop for chen distribution
}
