% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chen_rand.R
\name{chen_rand}
\alias{chen_rand}
\title{chen_rand}
\usage{
chen_rand(n, md, lambda, tau)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required}

\item{lambda}{shape parameter}

\item{tau}{a number that indicates the quantile}
}
\description{
random generation for the chen distribution.
}
\details{
The  reparameterized chen distribution has density:
\deqn{f(y;\lambda,\mu,\tau)=\dfrac{log(1-\tau)}{1-exp(\mu^\lambda)}\lambda y^{\lambda-1}exp\bigg[\dfrac{log(1-\tau)}{1-exp(\mu^\lambda]}[1-exp(y^\lambda)+y^\lambda \bigg]}
}
