% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chen_reg.R
\name{chen_reg}
\alias{chen_reg}
\title{chen_reg}
\usage{
chen_reg(data, formula, quantile = 0.5, link = "log")
}
\arguments{
\item{data}{data frame, list or environment (or object coercible by
as.data.frame to a data frame)
               containing the variables in the model.}

\item{formula}{an object of class "formula" (or one that can be coerced
to that class):
                 a symbolic description of the model to be fitted}

\item{quantile}{a number that indicates the quantile that you want to fit
the regression}

\item{link}{string that indicates the link function that you want to
fit the regression. Either
               "log" or "sqrt"}
}
\value{
chenReg returns an object of class `lm`
For more information on class `lm` type ?lm on your console
}
\description{
chen_reg is used to fit the chen regression, specified by giving the
formula, the quantile, and the link function.
}
\note{
You can specify all variables (except y) to be your covariates using `.`, you
can also add transformation like `log(x1)`
}
\examples{
library(chenreg)
chen_reg(data = simu[, -1], Y ~ V2 + V3, quantile = 0.5, link = "log")
chen_reg(data = simu[, -1], Y ~ ., quantile = 0.2, link = "log")

}
