% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chen_summary.R
\name{chen_summary}
\alias{chen_summary}
\title{chen_summary}
\usage{
chen_summary(z)
}
\arguments{
\item{z}{a model created by chen_reg.fit}
}
\description{
Create a summary of the model in the format of `data.frame`
}
\examples{

modell=chenReg::chen_reg.fit(Y ~.,data=simu[,-1],tau=0.5, link="log")
modell |>
chen_summary()
}
