% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.chenreg.R
\name{coef.chenreg}
\alias{coef.chenreg}
\title{coef.chenreg}
\usage{
\method{coef}{chenreg}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model coefficients is meaningful.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
coef.chenreg
}
