% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.chenreg.R
\name{fitted.chenreg}
\alias{fitted.chenreg}
\title{fitted.chenreg}
\usage{
\method{fitted}{chenreg}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model fitted values is meaningful.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
fitted.chenreg
}
