% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLike.chenreg.R
\name{logLike.chenreg}
\alias{logLike.chenreg}
\title{logLike.chenreg}
\usage{
logLike.chenreg(object, ...)
}
\arguments{
\item{object}{a fitted model created by `chen_reg`,
for which there exists the log-likelihood.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
logLike.chenreg
}
