% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.chenreg.R
\name{model.matrix.chenreg}
\alias{model.matrix.chenreg}
\title{model.matrix.chenreg}
\usage{
\method{model.matrix}{chenreg}(object, ...)
}
\arguments{
\item{object}{a fitted model created by `chen_reg`,
for which there exists the log-likelihood.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
model.matrix.chenreg
}
