% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.chenreg.R
\name{plot.chenreg}
\alias{plot.chenreg}
\title{plot.chenreg}
\usage{
\method{plot}{chenreg}(x, titles = TRUE, ...)
}
\arguments{
\item{x}{a model created by `chen_reg`}

\item{titles}{Title to appear above the plots}

\item{...}{further arguments passed to or from other methods.}
}
\description{
plot.chenreg
}
