% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.chenreg.R
\name{predict.chenreg}
\alias{predict.chenreg}
\title{predict.chenreg}
\usage{
\method{predict}{chenreg}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{a model created by `chen_reg`}

\item{newdata}{An optional data frame in which to look for variables
with which to predict. If omitted, the fitted values are used}

\item{...}{further arguments passed to or from other methods.}
}
\description{
predict.chenreg
}
