% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.autochen.R
\name{print.autochen}
\alias{print.autochen}
\title{print.autochen}
\usage{
\method{print}{autochen}(x, ...)
}
\arguments{
\item{x}{a model created by `auto_chen`}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print.autochen
}
