% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.chenreg.R
\name{print.summary.chenreg}
\alias{print.summary.chenreg}
\title{print.summary.chenreg}
\usage{
\method{print}{summary.chenreg}(x, digits = max(3, getOption("digits") - 3), ...)
}
\description{
print.summary.chenreg
}
