% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.chenreg.R
\name{residuals.chenreg}
\alias{residuals.chenreg}
\title{residuals.chenreg}
\usage{
\method{residuals}{chenreg}(object, ...)
}
\arguments{
\item{object}{a model created by `chen_reg`}

\item{...}{further arguments passed to or from other methods.}
}
\description{
residuals.chenreg
}
