% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.chenreg.R
\name{summary.chenreg}
\alias{summary.chenreg}
\title{summary.chenreg}
\usage{
\method{summary}{chenreg}(object, format = FALSE, ...)
}
\arguments{
\item{object}{a model created by `chen_reg`}

\item{...}{further arguments passed to or from other methods.}
}
\description{
summary.chenreg
}
