% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chen_envlp.R
\name{chen_envlp}
\alias{chen_envlp}
\title{chen_envlp}
\usage{
chen_envlp(z, b = 100)
}
\arguments{
\item{z}{a model created by `chen_reg`}

\item{b}{quantity of resamples}
}
\value{
Return the simulated envelope in class `ggplot`
}
\description{
Create a simulated envelope to see if the model fits the  Chen distribution well.
}
\note{
by default the quantity of resamples is 100
}
\examples{
chenreg::chen_reg(formula = Y ~., link = 'log', quantile = 0.5,
data = chenreg::simu[,-1]) |> chenreg::chen_envlp()

}
