% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chen_envlp.R
\name{chen_envlp}
\alias{chen_envlp}
\title{chen_envlp}
\usage{
chen_envlp(z, b = 100)
}
\arguments{
\item{z}{a model created by `chen_reg.fit`}

\item{b}{quantity of resamples}
}
\description{
Create a simulated envelope to see if the model fits the chen distribution well.
}
\note{
by default the quantity of resamples is 100
}
