% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chen_reg.fit.R
\name{chen_reg.fit}
\alias{chen_reg.fit}
\title{chen_reg.fit}
\usage{
chen_reg.fit(formula, data, tau, link)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class):
a symbolic description of the model to be fitted}

\item{data}{data frame, list or environment (or object coercible by as.data.frame to a data frame)
containing the variables in the model.}

\item{tau}{a number that indicates the quantile that you want to fit the regression}

\item{link}{string that indicates the link function that you want to fit the regression. Either
"log" or "sqrt"}
}
\description{
used to fit the  quantile regression
}
