% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.chenreg.R
\name{print.summary.chenreg}
\alias{print.summary.chenreg}
\title{print.summary.chenreg}
\usage{
\method{print}{summary.chenreg}(x, ...)
}
\arguments{
\item{x}{a model created by `chen_reg`}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print.summary.chenreg
}
