% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simu}
\alias{simu}
\title{Example Dataset to fit the regression
Contains five variables}
\format{
A dataframe with 300 rows and 4 variables
}
\source{
using the function from package
}
\usage{
simu
}
\description{
Example Dataset to fit the regression
Contains five variables
}
\keyword{datasets}
